app/py-service/on-chain-analysis/requirements.txt
requests==2.26.0
google-generativeai==0.1.0
solana==0.29.1
base58==2.1.1